python-fitbit/requirements.txt
alabaster==0.7.10
Babel==2.5.3
certifi==2018.1.18
chardet==3.0.4
CherryPy==3.8.2
click==6.7
coverage==3.7.1
docutils==0.14
fitbit==0.3.0
Flask==0.12.2
freezegun==0.3.10
funcsigs==1.0.2
httplib2==0.11.3
idna==2.6
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
mock==2.0.0
numpy==1.14.2
oauthlib==2.0.7
pandas==0.22.0
pbr==4.0.2
pluggy==0.3.1
py==1.5.3
Pygments==2.2.0
python-crontab==2.2.8
python-dateutil==2.7.2
pytz==2018.4
requests==2.18.4
requests-mock==1.4.0
requests-oauthlib==0.8.0
six==1.11.0
snowballstemmer==1.2.1
Sphinx==1.3.6
sphinx-rtd-theme==0.3.0
tox==2.1.1
urllib3==1.22
virtualenv==15.2.0
Werkzeug==0.14.1
